requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.1
cron-descriptor==1.4.5
Django==5.1.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
-e git+https://github.com/LogizonBuilds/glint_pm.git@b9a028273d224ffe804861298b320f28e1a9394a#egg=glint_pm
idna==3.10
kombu==5.4.2
phonenumbers==8.13.35
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.2
six==1.17.0
sparky_utils==1.1.7
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
